Almost finished, first pass version
Outside unexpected behaviours (error/duplicate handling) the only main feature missing is adding brand new sound banks to the sound bank list
