+++ Studio Sim/Resources/Audio/Sound Banks/SoundBank3.json
@@ -1,0 +1,37 @@
+[
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\pcm-32bit-44khz-mono.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\pcm-32bit-44khz-stereo.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\piano2.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\quietlaugh.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\bullettest.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 1,
+        "filePath": "Resources\\Audio\\creepymusic.wav",
+        "volume": 1.0
+    },
+    {
+        "audioType": 1,
+        "filePath": "Resources\\Audio\\partymusic.wav",
+        "volume": 1.0
+    }
+]
