+++ Studio Sim/Resources/Audio/Sound Banks/SoundBank4.json
@@ -1,0 +1,7 @@
+[
+    {
+        "audioType": 0,
+        "filePath": "Resources\\Audio\\bullettest.wav",
+        "volume": 0.375
+    }
+]
