--- Studio Sim/Resources/Audio/Sound Banks/soundFiles.json
@@ -1,22 +1,0 @@
-[
-    {
-        "audioType": 1,
-        "filePath": "Resources\\Audio\\partymusic.wav",
-        "volume": 0.2240000069141388
-    },
-    {
-        "audioType": 0,
-        "filePath": "Resources\\Audio\\piano2.wav",
-        "volume": 1.0
-    },
-    {
-        "audioType": 0,
-        "filePath": "Resources\\Audio\\quietlaugh.wav",
-        "volume": 1.0
-    },
-    {
-        "audioType": 0,
-        "filePath": "Resources\\Audio\\bullettest.wav",
-        "volume": 0.13500000536441803
-    }
-]
