--- Studio Sim/Resources/Audio/Sound Banks/soundFilesTestTest.json
@@ -1,17 +1,0 @@
-[
-    {
-        "audioType": 0,
-        "filePath": "Resources\\Audio\\pcm-32bit-44khz-mono.wav",
-        "volume": 0.33899998664855957
-    },
-    {
-        "audioType": 0,
-        "filePath": "Resources\\Audio\\quietlaugh.wav",
-        "volume": 1.0
-    },
-    {
-        "audioType": 1,
-        "filePath": "Resources\\Audio\\testfilepath.wav",
-        "volume": 0.6819999814033508
-    }
-]
