+++ Studio Sim/Studio Sim.vcxproj.filters
@@ -333,9 +333,6 @@
     <ClCompile Include="Utility\Rendering\DDSTextureLoader.cpp">
       <Filter>Source Files\Utility\Rendering</Filter>
     </ClCompile>
-    <ClCompile Include="Audio\AudioEditorNew.cpp">
-      <Filter>Source Files\Audio</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Application.h">
@@ -619,9 +616,6 @@
     </ClInclude>
     <ClInclude Include="Collisions\Rect2f.h">
       <Filter>Header Files\Utility\Helper</Filter>
-    </ClInclude>
-    <ClInclude Include="Audio\AudioEditorNew.h">
-      <Filter>Header Files\Audio</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
